Implement phieuxuatkho report.

+++ src/server/src/main/java/com/qlkh/server/handler/report/MaterialOutReportHandler.java
@@ -3,6 +3,7 @@ package com.qlkh.server.handler.report;
 import com.qlkh.core.client.action.report.MaterialOutReportAction;
 import com.qlkh.core.client.action.report.MaterialOutReportResult;
 import com.qlkh.core.client.report.MaterialReportBean;
+import com.qlkh.core.configuration.ConfigurationServerUtil;
 import com.qlkh.server.dao.SqlQueryDao;
 import com.qlkh.server.handler.core.AbstractHandler;
 import com.qlkh.server.servlet.ReportServlet;
@@ -37,10 +38,11 @@ import java.util.Map;
  */
 public class MaterialOutReportHandler extends AbstractHandler<MaterialOutReportAction, MaterialOutReportResult> {
 
+    private static final String REPORT_TEMPLATE_DIRECTORY = "report/template";
     private static final String JASPER_REPORT_FILE_NAME = "phieuxuatkho.jasper";
     private static final String OUTPUT_FILE_NAME = "phieuxuatkho";
+    private static final String PRINT_All_FILE_NAME = "printall.xls";
     private static final String OUTPUT_FILE_EXT = ".xls";
-    private static final String REPORT_SERVLET_URI = "/report/";
 
     @Autowired
     private SqlQueryDao sqlQueryDao;
@@ -52,8 +54,9 @@ public class MaterialOutReportHandler extends AbstractHandler<MaterialOutReportA
 
     @Override
     public MaterialOutReportResult execute(MaterialOutReportAction action, ExecutionContext context) throws DispatchException {
-        String reportFilePath = ServletUtils.getInstance().getRealPath(ReportServlet.REPORT_DIRECTORY, JASPER_REPORT_FILE_NAME);
-        String htmlFilePath = ServletUtils.getInstance().getRealPath(ReportServlet.REPORT_DIRECTORY, OUTPUT_FILE_NAME);
+        String reportFilePath = ServletUtils.getInstance().getRealPath(REPORT_TEMPLATE_DIRECTORY, JASPER_REPORT_FILE_NAME);
+        String printAllFilePath = ServletUtils.getInstance().getRealPath(REPORT_TEMPLATE_DIRECTORY, PRINT_All_FILE_NAME);
+        String outputFilePath = ServletUtils.getInstance().getRealPath(ReportServlet.REPORT_DIRECTORY, OUTPUT_FILE_NAME);
         List<String> ouputs = new ArrayList<String>();
         try {
             List<MaterialReportBean> materialReportBeans = sqlQueryDao.getMaterialOut(action.getRegex());
@@ -63,7 +66,7 @@ public class MaterialOutReportHandler extends AbstractHandler<MaterialOutReportA
                     data.put("stationName", materialReportBean.getStationName());
                     data.put("reason", materialReportBean.getReason());
                     data.put("personName", materialReportBean.getPersonName());
-                    data.put("code", materialReportBean.getCode());
+                    data.put("code", materialReportBean.getReportCode());
                     data.put("date", DateTimeUtils.dateTimeInVietnamese());
                     data.put("totalMoneyString", MoneyConverter
                             .transformNumber(String.valueOf(materialReportBean.getMoney().intValue())));
@@ -72,7 +75,7 @@ public class MaterialOutReportHandler extends AbstractHandler<MaterialOutReportA
                     beans.add(materialReportBean);
                     for (int i = 1; i <= 3; i++) {
                         data.put("reportName", "Liên " + i);
-                        String fileOutputPath = htmlFilePath + materialReportBean.getCode() + i + OUTPUT_FILE_EXT;
+                        String fileOutputPath = outputFilePath + materialReportBean.getCode() + i + OUTPUT_FILE_EXT;
                         JasperPrint jasperPrint = JasperFillManager.fillReport(reportFilePath, data,
                                 new JRBeanCollectionDataSource(beans));
                         ReportExporter.exportReportXls(jasperPrint, fileOutputPath);
@@ -86,26 +89,39 @@ public class MaterialOutReportHandler extends AbstractHandler<MaterialOutReportA
             e.printStackTrace();
         }
 
+        String reportUrl = null;
+
         try {
-            HSSFWorkbook reportWorkbook = new HSSFWorkbook();
-            for (String output : ouputs) {
-                File file = new File(output);
-                FileInputStream input = new FileInputStream(file);
-                HSSFWorkbook tempWorkbook = new HSSFWorkbook(input);
-                if (tempWorkbook.getNumberOfSheets() > 0) {
-                    HSSFSheet copySheet = tempWorkbook.getSheetAt(0);
-                    HSSFSheet newSheet = reportWorkbook.createSheet();
-                    ExcelUtil.copySheets(newSheet, copySheet);
+            if (CollectionUtils.isNotEmpty(ouputs)) {
+                File printAllFile = new File(printAllFilePath);
+                FileInputStream printAllFileInputStream = new FileInputStream(printAllFile);
+                HSSFWorkbook reportWorkbook = new HSSFWorkbook(printAllFileInputStream);
+                for (String output : ouputs) {
+                    File file = new File(output);
+                    FileInputStream input = new FileInputStream(file);
+                    HSSFWorkbook tempWorkbook = new HSSFWorkbook(input);
+                    if (tempWorkbook.getNumberOfSheets() > 0) {
+                        HSSFSheet copySheet = tempWorkbook.getSheetAt(0);
+                        HSSFSheet newSheet = reportWorkbook.createSheet();
+                        ExcelUtil.copySheets(newSheet, copySheet);
+                    }
+                    input.close();
+                    FileSystemUtils.deleteRecursively(file);
                 }
-                input.close();
-                FileSystemUtils.deleteRecursively(file);
+                Util.writeToFile(outputFilePath + OUTPUT_FILE_EXT, reportWorkbook);
+                printAllFileInputStream.close();
+                reportUrl = new StringBuilder().append(ConfigurationServerUtil.getServerBaseUrl())
+                        .append(ConfigurationServerUtil.getConfiguration().serverServletRootPath())
+                        .append(ReportServlet.REPORT_SERVLET_URI)
+                        .append(ReportServlet.REPORT_FILENAME_PARAMETER)
+                        .append("=")
+                        .append(OUTPUT_FILE_NAME + OUTPUT_FILE_EXT).toString();
             }
-            Util.writeToFile(htmlFilePath + OUTPUT_FILE_EXT, reportWorkbook);
         } catch (FileNotFoundException e) {
             e.printStackTrace();
         } catch (IOException e) {
             e.printStackTrace();
         }
-        return new MaterialOutReportResult(ouputs);
+        return new MaterialOutReportResult(reportUrl);
     }
 }
